refactor(Constraint Authoring): Extract EditConstraintRemovalCriteriaComponent
#1182
