+++ src/assets/wise5/authoringTool/constraint/constraint-authoring/constraint-authoring.component.scss
@@ -13,7 +13,3 @@
   margin-bottom: 10px;
   padding: 10px 20px 10px 20px;
 }
-
-.red {
-  color: red;
-}
