+++ src/assets/wise5/authoringTool/constraint/edit-constraint-removal-criteria/edit-constraint-removal-criteria.component.scss
@@ -1,0 +1,8 @@
+.common-width {
+  width: 400px;
+}
+
+.red {
+  color: red;
+}
+
